nrf52_bsim: Move CMSIS functions to cmsis.c
Only a few CMSIS functions are added for now
