+++ boards/posix/nrf52_bsim/irq_handler.c
@@ -13,7 +13,6 @@
 #include "kswap.h"
 #include "irq_ctrl.h"
 #include "posix_core.h"
-#include "posix_board_if.h"
 #include "board_soc.h"
 #include "sw_isr_table.h"
 #include "soc.h"
@@ -365,63 +364,6 @@ void posix_irq_offload(void (*routine)(void *), void *parameter)
 	posix_irq_disable(OFFLOAD_SW_IRQ);
 }
 #endif
-
-/*
- *  Replacement for ARMs NVIC functions()
- */
-void NVIC_SetPendingIRQ(IRQn_Type IRQn)
-{
-	hw_irq_ctrl_raise_im_from_sw(IRQn);
-}
-
-void NVIC_ClearPendingIRQ(IRQn_Type IRQn)
-{
-	hw_irq_ctrl_clear_irq(IRQn);
-}
-
-void NVIC_DisableIRQ(IRQn_Type IRQn)
-{
-	hw_irq_ctrl_disable_irq(IRQn);
-}
-
-void NVIC_EnableIRQ(IRQn_Type IRQn)
-{
-	hw_irq_ctrl_enable_irq(IRQn);
-}
-
-void NVIC_SetPriority(IRQn_Type IRQn, uint32_t priority)
-{
-	hw_irq_ctrl_prio_set(IRQn, priority);
-}
-
-uint32_t NVIC_GetPriority(IRQn_Type IRQn)
-{
-	return hw_irq_ctrl_get_prio(IRQn);
-}
-
-void NVIC_SystemReset(void)
-{
-	posix_print_warning("%s called. Exiting\n", __func__);
-	posix_exit(1);
-}
-
-/*
- * Replacements for some other CMSIS functions
- */
-void __enable_irq(void)
-{
-	hw_irq_ctrl_change_lock(true);
-}
-
-void __disable_irq(void)
-{
-	hw_irq_ctrl_change_lock(false);
-}
-
-uint32_t __get_PRIMASK(void)
-{
-	return hw_irq_ctrl_get_current_lock();
-}
 
 /*
  * Very simple model of the WFE and SEV ARM instructions
